requirements.txt
scikit-learn==0.23.2
azureml-mlflow==1.18.0.post1
matplotlib==3.3.3